Updated Dashboard
Updated the dashboard to it actually sends a file, not just the file name, had to update a few on the fileapp (fileAPI, and R2Lib) to accommodate.  The sent file is now stored in the computer's temp directory,.  Works for stand along environment, will have to change to S3 when we put on the cloud.  Still need to work on the display and replay screen.